backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
requests>=2.31.0
python-dotenv>=1.0.0
httpx>=0.24.0
supabase>=2.0.0
google-genai>=0.1.0
openai>=1.0.0
boto3>=1.34.0
apscheduler>=3.10.0
Pillow>=10.0.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
